STYLE: Rename "MyStandardResampler" directory to "DefaultResampler"
