+++ Components/Resamplers/DefaultResampler/elxMyStandardResampler.cxx
