+++ Components/Resamplers/DefaultResampler/elxMyStandardResampler.h
