+++ Components/Resamplers/DefaultResampler/elxMyStandardResampler.hxx
